Parse out arbitrary dates
This will parse out dates in the format of "mon dd", the month being a 3
letter month, and the date being just the day.

due mar 10
due sep 25
due oct 2

it will also "intelligently" guess the year, based upon which is the
shorter duration from time.Now(). for instance, if today's date is dec
25 and you schedule an event for jan 2, it will schedule it for the
following year.