Global repo file as the default
This patch implments the new workflow mentioned in #31, in which
the program still tries to find a repo file in the working directory
first, and it tries the global one in the home directory if no one
is found.

In this patch, the condition check of the existence of the repo file
is done in `todolist/file_store.go` at the very early step before
the input routing.