Fix listing overdue tasks
This bug was because the regex we're using to determine filtering by a
date is expecting to see `due <something>`.  `overdue` was part of that
switch statement.

I pushed `overdue` up to be part of the "special cases" (with "agenda"),
so that it will run the `filterOverdue` logic before going into the big
switch statement.

I found that testing this in the real-world was hard.

closes #101