Refactor the formatter to an interface and rename it
This change refactors the formatter to be an interface called Printer.
`NewApp` will use the functionality of `ScreenPrinter`,  which is what I
refactored the old `Formatter` class into.

There is a new class the implements the `Printer` interface called
`MemoryPrinter`, which will simply store the `groups` as they were
passed to it.  This will allow tests in `app_test.go` to become much
more powerful, since `App` no longer needs to be tightly coupled to
printing to the screen.