change how background sync works
previously we were relying on the os-dependent `syscall.SysProcAttr` to detach the ultralist sync call from the parent.

However, the windows-specific `SysProcAttr` struct does not include `Setpgid` and `Pgid` attributes, and therefore I could not compile ultralist for windows.

This is now solved using the os-agnostic `os.StartProcess`, followed by the `Release` command.