Update the sync command (#211)
* Update the sync command

Previously, setting up a list to sync with ultralist.io was fairly convoluted.  Bot the `init` and `sync` commands were dependent on a list's sync state, and the behavior of these commands would change based upon that.

**Previous flow:**

* `ultralist init` - Create a new list and optionally, sync it with Ultralist.io.
* `ultralist sync` - Depending if a list is synced or not, it does multiple things:
  * If not synced, sets up the local list to sync to ultralist.io.
  * If is already synced, pull remote changes to local, and push any local changes to remote.

This PR simplifies the commands by adding a couple of flags to the `sync` command.  The result is that each command has one job to do, instead of many dependent on state.  It's simpler to understand as well.

**With this PR:**

* `ultralist init` - makes this command do just one thing - initialize a list.  It does not handle syncing a list as well.
* `ultralist sync --setup` - sets up a local list to sync with ultralist.io, or pulls a list from ultralist.io and replaces what's local.
* `ultralist sync --unsync` - stops a local list from syncing with ultralist.io.
* `ultralist sync` - just handles the actual list syncing between local and ultralist.io.

* allow a list to be synced before a .todos.json file exists

also, refactor file_store so it is less rigid.

* Clean up FileStore more, update tests

Co-authored-by: Grant Ammons <gammons@gmail.com>