Fix bug with syncing todos
You must have a synced list for this bug to be replicable.

If you have a `.todos.json` file in your home dir, and add a todo to it
in a subdirectory, it would end up creating a `.todos.json` file in the
subdirectory, rather than adding it to the `.todos.json` in the home
dir.