WIP on ultralist 2.0
This is a WIP on v2.0 of ultralist.

The main goal of this is to bring a higher level of professionalism and
structure to the project as a whole.   This will be achieved by:

1. **Splitting concerns into packages.**  There will be a core
`ultralist` pacakge, which will implement the main business logic around
Todo lists, todos, recurrence, etc.

2. **Eliminate dependence on CLI.**  Ultralist will eventually have a
more tui-like user interface, and there are a lot of places in the app
that have a reliance on ultralist being a CLI only.  Separate the
concerns of CLI interaction into a `cli` package to prepare for that.

3. **Proper use of errors**.  Improve error handling across all packages.

4. **Prepare for a re-think of the backend + syncing.**  Ultralist
should use a store like CouchDB for syncing.  Prepare the app for that
eventuality.